refactor: moves library blocks code into blocks modules
