+++ openedx/core/djangoapps/content_libraries/rest_api/libraries.py
@@ -69,11 +69,9 @@ import logging
 from django.conf import settings
 from django.contrib.auth import authenticate, get_user_model, login
 from django.contrib.auth.models import Group
-from django.core.exceptions import ObjectDoesNotExist
 from django.db.transaction import atomic, non_atomic_requests
-from django.http import Http404, HttpResponse, HttpResponseBadRequest, JsonResponse, StreamingHttpResponse
+from django.http import Http404, HttpResponseBadRequest, JsonResponse
 from django.shortcuts import get_object_or_404
-from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.translation import gettext as _
 from django.views.decorators.clickjacking import xframe_options_exempt
@@ -85,14 +83,12 @@ from pylti1p3.exception import LtiException, OIDCException
 
 import edx_api_doc_tools as apidocs
 from opaque_keys.edx.locator import LibraryLocatorV2, LibraryUsageLocatorV2
-from openedx_learning.api import authoring
 from organizations.api import ensure_organization
 from organizations.exceptions import InvalidOrganizationException
 from organizations.models import Organization
 from rest_framework import status
 from rest_framework.exceptions import NotFound, PermissionDenied, ValidationError
 from rest_framework.generics import GenericAPIView
-from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.viewsets import GenericViewSet
@@ -110,13 +106,9 @@ from openedx.core.djangoapps.content_libraries.rest_api.serializers import (
     ContentLibraryPermissionLevelSerializer,
     ContentLibraryPermissionSerializer,
     ContentLibraryUpdateSerializer,
-    ContentLibraryComponentCollectionsUpdateSerializer,
     LibraryXBlockCreationSerializer,
     LibraryXBlockMetadataSerializer,
     LibraryXBlockTypeSerializer,
-    LibraryXBlockOlxSerializer,
-    LibraryXBlockStaticFileSerializer,
-    LibraryXBlockStaticFilesSerializer,
     ContentLibraryAddPermissionByEmailSerializer,
     LibraryPasteClipboardSerializer,
 )
@@ -124,7 +116,6 @@ import openedx.core.djangoapps.site_configuration.helpers as configuration_helpe
 from openedx.core.lib.api.view_utils import view_auth_classes
 from openedx.core.djangoapps.safe_sessions.middleware import mark_user_change_as_expected
 from openedx.core.djangoapps.xblock import api as xblock_api
-from openedx.core.types.http import RestRequest
 
 from .utils import convert_exceptions
 from ..models import ContentLibrary, LtiGradedResource, LtiProfile
@@ -632,212 +623,6 @@ class LibraryBlockView(APIView):
         return Response({})
 
 
-@view_auth_classes()
-class LibraryBlockRestore(APIView):
-    """
-    View to restore soft-deleted library xblocks.
-    """
-    @convert_exceptions
-    def post(self, request, usage_key_str) -> Response:
-        """
-        Restores a soft-deleted library block that belongs to a Content Library
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
-        api.restore_library_block(key)
-        return Response(None, status=status.HTTP_204_NO_CONTENT)
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockCollectionsView(APIView):
-    """
-    View to set collections for a component.
-    """
-    @convert_exceptions
-    def patch(self, request: RestRequest, usage_key_str) -> Response:
-        """
-        Sets Collections for a Component.
-
-        Collection and Components must all be part of the given library/learning package.
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        content_library = api.require_permission_for_library_key(
-            key.lib_key,
-            request.user,
-            permissions.CAN_EDIT_THIS_CONTENT_LIBRARY
-        )
-        component = api.get_component_from_usage_key(key)
-        serializer = ContentLibraryComponentCollectionsUpdateSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-
-        collection_keys = serializer.validated_data['collection_keys']
-        api.set_library_component_collections(
-            library_key=key.lib_key,
-            component=component,
-            collection_keys=collection_keys,
-            created_by=request.user.id,
-            content_library=content_library,
-        )
-
-        return Response({'count': len(collection_keys)})
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockLtiUrlView(APIView):
-    """
-    Views to generate LTI URL for existing XBlocks in a content library.
-
-    Returns 404 in case the block not found by the given key.
-    """
-    @convert_exceptions
-    def get(self, request, usage_key_str):
-        """
-        Get the LTI launch URL for the XBlock.
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
-
-        # Get the block to validate its existence
-        api.get_library_block(key)
-        lti_login_url = f"{reverse('content_libraries:lti-launch')}?id={key}"
-        return Response({"lti_url": lti_login_url})
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockOlxView(APIView):
-    """
-    Views to work with an existing XBlock's OLX
-    """
-    @convert_exceptions
-    def get(self, request, usage_key_str):
-        """
-        DEPRECATED. Use get_block_olx_view() in xblock REST-API.
-        Can be removed post-Teak.
-
-        Get the block's OLX
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
-        xml_str = xblock_api.get_block_draft_olx(key)
-        return Response(LibraryXBlockOlxSerializer({"olx": xml_str}).data)
-
-    @convert_exceptions
-    def post(self, request, usage_key_str):
-        """
-        Replace the block's OLX.
-
-        This API is only meant for use by developers or API client applications.
-        Very little validation is done.
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY)
-        serializer = LibraryXBlockOlxSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        new_olx_str = serializer.validated_data["olx"]
-        try:
-            version_num = api.set_library_block_olx(key, new_olx_str).version_num
-        except ValueError as err:
-            raise ValidationError(detail=str(err))  # lint-amnesty, pylint: disable=raise-missing-from
-        return Response(LibraryXBlockOlxSerializer({"olx": new_olx_str, "version_num": version_num}).data)
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockAssetListView(APIView):
-    """
-    Views to list an existing XBlock's static asset files
-    """
-    @convert_exceptions
-    def get(self, request, usage_key_str):
-        """
-        List the static asset files belonging to this block.
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
-        files = api.get_library_block_static_asset_files(key)
-        return Response(LibraryXBlockStaticFilesSerializer({"files": files}).data)
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockAssetView(APIView):
-    """
-    Views to work with an existing XBlock's static asset files
-    """
-    parser_classes = (MultiPartParser, )
-
-    @convert_exceptions
-    def get(self, request, usage_key_str, file_path):
-        """
-        Get a static asset file belonging to this block.
-        """
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(key.lib_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY)
-        files = api.get_library_block_static_asset_files(key)
-        for f in files:
-            if f.path == file_path:
-                return Response(LibraryXBlockStaticFileSerializer(f).data)
-        raise NotFound
-
-    @convert_exceptions
-    def put(self, request, usage_key_str, file_path):
-        """
-        Replace a static asset file belonging to this block.
-        """
-        file_path = file_path.replace(" ", "_")  # Messes up url/name correspondence due to URL encoding.
-        usage_key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(
-            usage_key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY,
-        )
-        file_wrapper = request.data['content']
-        if file_wrapper.size > 20 * 1024 * 1024:  # > 20 MiB
-            # TODO: This check was written when V2 Libraries were backed by the Blockstore micro-service.
-            #       Now that we're on Learning Core, do we still need it? Here's the original comment:
-            #         In the future, we need a way to use file_wrapper.chunks() to read
-            #         the file in chunks and stream that to Blockstore, but Blockstore
-            #         currently lacks an API for streaming file uploads.
-            #       Ref:  https://github.com/openedx/edx-platform/issues/34737
-            raise ValidationError("File too big")
-        file_content = file_wrapper.read()
-        try:
-            result = api.add_library_block_static_asset_file(usage_key, file_path, file_content, request.user)
-        except ValueError:
-            raise ValidationError("Invalid file path")  # lint-amnesty, pylint: disable=raise-missing-from
-        return Response(LibraryXBlockStaticFileSerializer(result).data)
-
-    @convert_exceptions
-    def delete(self, request, usage_key_str, file_path):
-        """
-        Delete a static asset file belonging to this block.
-        """
-        usage_key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.require_permission_for_library_key(
-            usage_key.lib_key, request.user, permissions.CAN_EDIT_THIS_CONTENT_LIBRARY,
-        )
-        try:
-            api.delete_library_block_static_asset_file(usage_key, file_path, request.user)
-        except ValueError:
-            raise ValidationError("Invalid file path")  # lint-amnesty, pylint: disable=raise-missing-from
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-@method_decorator(non_atomic_requests, name="dispatch")
-@view_auth_classes()
-class LibraryBlockPublishView(APIView):
-    """
-    Commit/publish all of the draft changes made to the component.
-    """
-
-    @convert_exceptions
-    def post(self, request, usage_key_str):
-        key = LibraryUsageLocatorV2.from_string(usage_key_str)
-        api.publish_component_changes(key, request.user)
-        return Response({})
-
-
 @method_decorator(non_atomic_requests, name="dispatch")
 @view_auth_classes()
 class LibraryImportTaskViewSet(GenericViewSet):
@@ -859,7 +644,7 @@ class LibraryImportTaskViewSet(GenericViewSet):
             request.user,
             permissions.CAN_VIEW_THIS_CONTENT_LIBRARY
         )
-        queryset = api.ContentLibrary.objects.get_by_key(library_key).import_tasks
+        queryset = ContentLibrary.objects.get_by_key(library_key).import_tasks
         result = ContentLibraryBlockImportTaskSerializer(queryset, many=True).data
 
         return self.get_paginated_response(
@@ -1175,105 +960,3 @@ class LtiToolJwksView(LtiToolView):
         Return the JWKS.
         """
         return JsonResponse(self.lti_tool_config.get_jwks(), safe=False)
-
-
-def get_component_version_asset(request, component_version_uuid, asset_path):
-    """
-    Serves static assets associated with particular Component versions.
-
-    Important notes:
-    * This is meant for Studio/authoring use ONLY. It requires read access to
-      the content library.
-    * It uses the UUID because that's easier to parse than the key field (which
-      could be part of an OpaqueKey, but could also be almost anything else).
-    * This is not very performant, and we still want to use the X-Accel-Redirect
-      method for serving LMS traffic in the longer term (and probably Studio
-      eventually).
-    """
-    try:
-        component_version = authoring.get_component_version_by_uuid(
-            component_version_uuid
-        )
-    except ObjectDoesNotExist as exc:
-        raise Http404() from exc
-
-    # Permissions check...
-    learning_package = component_version.component.learning_package
-    library_key = LibraryLocatorV2.from_string(learning_package.key)
-    api.require_permission_for_library_key(
-        library_key, request.user, permissions.CAN_VIEW_THIS_CONTENT_LIBRARY,
-    )
-
-    # We already have logic for getting the correct content and generating the
-    # proper headers in Learning Core, but the response generated here is an
-    # X-Accel-Redirect and lacks the actual content. We eventually want to use
-    # this response in conjunction with a media reverse proxy (Caddy or Nginx),
-    # but in the short term we're just going to remove the redirect and stream
-    # the content directly.
-    redirect_response = authoring.get_redirect_response_for_component_asset(
-        component_version_uuid,
-        asset_path,
-        public=False,
-    )
-
-    # If there was any error, we return that response because it will have the
-    # correct headers set and won't have any X-Accel-Redirect header set.
-    if redirect_response.status_code != 200:
-        return redirect_response
-
-    # If we got here, we know that the asset exists and it's okay to download.
-    cv_content = component_version.componentversioncontent_set.get(key=asset_path)
-    content = cv_content.content
-
-    # Delete the re-direct part of the response headers. We'll copy the rest.
-    headers = redirect_response.headers
-    headers.pop('X-Accel-Redirect')
-
-    # We need to set the content size header manually because this is a
-    # streaming response. It's not included in the redirect headers because it's
-    # not needed there (the reverse-proxy would have direct access to the file).
-    headers['Content-Length'] = content.size
-
-    if request.method == "HEAD":
-        return HttpResponse(headers=headers)
-
-    # Otherwise it's going to be a GET response. We don't support response
-    # offsets or anything fancy, because we don't expect to run this view at
-    # LMS-scale.
-    return StreamingHttpResponse(
-        content.read_file().chunks(),
-        headers=redirect_response.headers,
-    )
-
-
-@view_auth_classes()
-class LibraryComponentAssetView(APIView):
-    """
-    Serves static assets associated with particular Component versions.
-    """
-    @convert_exceptions
-    def get(self, request, component_version_uuid, asset_path):
-        """
-        GET API for fetching static asset for given component_version_uuid.
-        """
-        return get_component_version_asset(request, component_version_uuid, asset_path)
-
-
-@view_auth_classes()
-class LibraryComponentDraftAssetView(APIView):
-    """
-    Serves the draft version of static assets associated with a Library Component.
-
-    See `get_component_version_asset` for more details
-    """
-    @convert_exceptions
-    def get(self, request, usage_key, asset_path):
-        """
-        Fetches component_version_uuid for given usage_key and returns component asset.
-        """
-        try:
-            component_version_uuid = api.get_component_from_usage_key(usage_key).versioning.draft.uuid
-        except ObjectDoesNotExist as exc:
-            raise Http404() from exc
-
-        return get_component_version_asset(request, component_version_uuid, asset_path)
